Rename _cse.py to cse.py to match project conventions
cvxpy uses bare acronyms as module names (psd.py, soc_canon.py,
scs_conif.py, dcp2cone/, dgp2dcp/, dnlp2smooth/, etc.) and does not
prefix internal modules with an underscore. cse (Common Subexpression
Elimination) is descriptive enough on its own.
