+++ cvxpy/reductions/dcp2cone/dcp2cone.py
@@ -21,8 +21,8 @@ from cvxpy.atoms.quad_over_lin import quad_over_lin
 from cvxpy.expressions import cvxtypes
 from cvxpy.expressions.expression import Expression
 from cvxpy.problems.objective import Minimize
-from cvxpy.reductions._cse import UncacheableError, expr_key
 from cvxpy.reductions.canonicalization import Canonicalization
+from cvxpy.reductions.cse import UncacheableError, expr_key
 from cvxpy.reductions.dcp2cone.canonicalizers import CANON_METHODS as cone_canon_methods
 from cvxpy.reductions.dcp2cone.canonicalizers.quad import QUAD_CANON_METHODS as quad_canon_methods
 from cvxpy.reductions.inverse_data import InverseData
